pyproject.toml
[project]
name = "alexandria-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.2",
    "channels-redis>=4.2.1",
    "channels[daphne]>=4.2.2",
    "django>=5.2.1",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "redis>=6.1.0",
]
